requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-settings==2.2.1
httpx==0.27.0
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.9
pytest==8.1.1
pytest-asyncio==0.23.6
anyio==4.3.0
bcrypt>=4.0
itsdangerous==2.1.2
python-jose[cryptography]==3.3.0
scikit-learn>=1.4.0
anthropic>=0.30.0
pandas>=2.2.0
spacy>=3.8.0
tqdm>=4.66.0
